+++ src/rest.cpp
@@ -43,7 +43,7 @@ static const struct {
 };
 
 struct CCoin {
-    uint32_t nHeight;
+    int32_t nHeight;
     CTxOut out;
 
     CCoin() : nHeight(0) {}
@@ -666,12 +666,12 @@ static bool rest_getutxos(Config &config, HTTPRequest *req,
             UniValue utxos(UniValue::VARR);
             for (const CCoin &coin : outs) {
                 UniValue utxo(UniValue::VOBJ);
-                utxo.push_back(Pair("height", int32_t(coin.nHeight)));
+                utxo.push_back(Pair("height", coin.nHeight));
                 utxo.push_back(Pair("value", ValueFromAmount(coin.out.nValue)));
 
                 // include the script in a json output
                 UniValue o(UniValue::VOBJ);
-                int height = (coin.nHeight == MEMPOOL_HEIGHT) ? (chainActive.Height() + 1) : coin.nHeight;
+                int32_t height = (coin.nHeight == MEMPOOL_HEIGHT) ? (chainActive.Height() + 1) : coin.nHeight;
                 ScriptPubKeyToUniv(coin.out.scriptPubKey, true, IsGenesisEnabled(config, height), o);
                 utxo.push_back(Pair("scriptPubKey", o));
                 utxos.push_back(utxo);
