+++ src/test/config_tests.cpp
@@ -87,7 +87,7 @@ BOOST_AUTO_TEST_CASE(max_block_size_related_defaults) {
 
     // Make up genesis activation parameters 
     // - Genesis will be activated at block height 100
-    uint64_t heightActivateGenesis   = 100;
+    int32_t heightActivateGenesis   = 100;
     config.SetGenesisActivationHeight(heightActivateGenesis);
 
     // Providing defaults should not override anything
